refactor(templates): rename AI Chat shell example to "Side Nav"
Rename the slug shell-ai-chat → shell-side-nav, set the display name to
"Side Nav", and update the docsite gallery component map. The example now
reads as a nav-pattern showcase (side-nav app shell) rather than a themed
app.
