+++ apps/docsite/src/components/templateComponents.ts
@@ -117,8 +117,9 @@ export const TEMPLATE_COMPONENTS: Record<
   'side-gallery': lazy(
     () => import('../../../../packages/cli/templates/pages/side-gallery/page'),
   ),
-  'shell-ai-chat': lazy(
-    () => import('../../../../packages/cli/templates/pages/shell-ai-chat/page'),
+  'shell-side-nav': lazy(
+    () =>
+      import('../../../../packages/cli/templates/pages/shell-side-nav/page'),
   ),
   table: lazy(
     () => import('../../../../packages/cli/templates/pages/table/page'),
