--- packages/cli/templates/pages/shell-ai-chat/template.doc.mjs
@@ -1,12 +1,0 @@
-// Copyright (c) Meta Platforms, Inc. and affiliates.
-
-/** @type {import('../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'page',
-  name: 'AI Chat Shell',
-  displayName: 'AI Chat Shell',
-  description:
-    'An AI chat shell with a collapsible, resizable SideNav — new chat, search, library, and workspace-grouped conversations with status dots — over static grey-card conversation and composer placeholders.',
-  isReady: true,
-  category: 'Shell - Left Sidebar',
-};
