+++ packages/cli/templates/pages/shell-side-nav/template.doc.mjs
@@ -1,0 +1,12 @@
+// Copyright (c) Meta Platforms, Inc. and affiliates.
+
+/** @type {import('../../../../core/src/docs-types').TemplateDoc} */
+export const doc = {
+  type: 'page',
+  name: 'Side Nav',
+  displayName: 'Side Nav',
+  description:
+    'A side-nav app shell with a collapsible, resizable SideNav — new chat, search, library, and workspace-grouped conversations with status dots — over static grey-card conversation and composer placeholders.',
+  isReady: true,
+  category: 'Shell - Left Sidebar',
+};
